Fixed dwarf4 line info reader.
The line information reader was adding lines to the line info matrix
for every dwarf4 instruction. Not all dwarf4 instructions are supposed
to add table entries. This resulted in intermediate steps being recorded
in the table so there would be multiple entries for the same address
at different line numbers. Lookup took the first entry found which was
not always the correct one. Intermediate steps are no longer added to the table.