Possible solution to Issue #49:
The code appears to remove blocks based on the group index instead
of the block index, while the sampler assumes that the given index
is a block index. This will require extensive testing to be sure
because the group index has been used since groups were implemented
so I could be missing something.