Fixes Issue #51 which was generating incorrect assembly to collect the index vector from some vector instructions.
The test application, TestDfft, had many scatter and gather cases
in the same block. Before the change, some instrumented correctly
while others did not. Now, they all instrument correctly. Two changes
were required. The first was that the B bit was always set to 1, but
for 64-bit, the higher GPRs (base registers) had to have this bit set
to 0. The other change was a bug, also related to the high GPR base
registers which always put a 1 when sometimes it needed to be a 0.