This is a large commit to get multi-image, multi-threaded applications working with the AddressStreamIntercept tool.
A minor changes were made to all the tools: `tool_dynamic_init`
initializes the dynamic instrumentation for each image. Dynamic
instrumentation uses the image sequence number to know which image
the dynamic instrumentation belongs to. `tool_dynamic_init` is not
turned off before `tool_image_init`, which can cause an issue with
multi-image applications (since shared libraries instrument every
function with both of these instrumentation functions). Originally,
`tool_dynamic_init` would control the image sequence number, but with
multiple entries into `tool_dynamic_init`, we had to use the Dynamic
Instrumentation object to keep track of the image sequence number. It
uses a global to do the count, and keeps track of which "dynamics"
have already been initialized.

In an attempt to serialize entries into `tool_dynamic_init` and
`tool_image_init` in AddressStreamIntercept, we surrounded both with
the same lock. We could not use a mutex, because `tool_image_init`
turns off instrumentation, and this caused deadlock. Instead, we
used a ReadWrite Lock. This makes sure that the two functions are
never executed simultaneously. (Note: This may not be necessary anymore,
but I'm not really sure...better to keep it in since it really doesn't
do any harm and not having it could do harm).

The biggest issue with the AddressStreamIntercept tool is that it
used the DataManager, which did not initialize data in any
particular order. The first thread/image would always be initialize
first, but after that, threads could come first or images could
spawn their own threads. This made it exceedingly difficult to know
how to "initialize" handlers and buffers. Handlers and buffers are
all shared by images, but each thread has their own. This worked
fine with multi-threaded, single image applications, since each thread
would just initialize their own handler/buffer. This also worked
for single-threaded, multi-image because new images would just
link to the first image. But, with multi-threaded, multi-image apps,
these assumptions fall short (again...any image can spawn threads).
To fix this, we had to modify the Data Manager to ALWAYS initialize
data structures for a new thread to the first image FIRST. Then, it
can initialize the data structures for the new thread to the other
images in any order. Then, while initializing handlers/buffers, if
it is the first image, we know to create a new one. If it is not,
then it can link to the first image (since it will have already been
created!).

The above change actually solved a second issue we had with images
sharing data structures. There was a memory leak because image
data could be deleted in any order. Now, we only delete it if it is
the first image.

The last issue was likely an issue for all multi-threaded applications
that just happened to manifest itself in our tests because of the
Data Manager "Refresh" function, which is not used in single-image
applications. The bug occurred while trying to process a buffer that
was seemingly empty (the image id was 0, which implies an empty entry).
This can happen with multiple threads and sampling for one of two
reasons (although I wouldn't rule out any others):
1) Sampling is turned off by another thread while the current thread is
in the middle of a block. The thread would have started filling out
entries from the first memops, but sampling would turn off and then it
would not be able to fill out any more entries.
2) Sampling is off during "data collection," so there are no entries
recorded. But then, while it is going to "process" (really "skip")
those addresses, sampling is turned back on by another thread. When
the current thread tries to "Process" these new entries (since sampling
is on again), then it finds empty entries.
The fix for this is not perfect. Unfortunately, we cannot give each
thread its own sampler, since threads share executed code. It is
also impossible to control threads while they are not in instrumentation
code. The best we can do is have them stop and continue again. The
solution we came up with for this issue is to just "skip" addresses
that are empty. The skip is passed to handlers so they can update
their stats accordingly.