Changed two checks that were breaking when they shouldn't have
Changed a check centered around PT_LOAD where originally the 3rd time
it was encountered we would throw an error assuming that only 2 PT_LOAD
sections exist when there can be more.
Second check was comparing the bloom filter to bitmask. Bit mask was
orignally 32 bits and we were doing a << by over 31 bits which should
have resulted in 0 but instead wrapped the bits around. So we sized
things up to prevent the overflow making bit mask 64 bits when
previously it was 32