PEBIL Issue #189: Wrote a prototype for the Ariel Frontend. To use the frontend, sst-core and sst-elements need to be installed and PEBIL must be configured with them.
The prototype only includes start and end instructions, read, write,
and exit commands. The instruction pointer is the memory sequence,
the address is the actual address, the thread is always 0, and the
rest of the Ariel Instruction information is hardcoded for now.

Prototype only works on serial applications at the moment.